refactor ActsTrackingGeometry to individual class
