+++ offline/packages/trackreco/ActsEvaluator.h
@@ -4,7 +4,7 @@
 #include <fun4all/SubsysReco.h>
 #include <trackbase/TrkrDefs.h>
 
-#include "PHActsSourceLinks.h"
+#include "ActsTrackingGeometry.h"
 #include "ActsTrack.h"
 
 #include <Acts/Utilities/Helpers.hpp>
