+++ offline/packages/trackreco/PHActsSourceLinks.h
@@ -24,6 +24,8 @@
 #include <ActsExamples/EventData/TrkrClusterSourceLink.hpp>
 #include <ActsExamples/Plugins/BField/BFieldOptions.hpp>
 
+#include "ActsTrackingGeometry.h"
+
 class PHCompositeNode;
 class TrkrClusterContainer;
 class TrkrCluster;
@@ -45,33 +47,6 @@ namespace Acts
 
 using Surface = std::shared_ptr<const Acts::Surface>;
 using SourceLink = ActsExamples::TrkrClusterSourceLink;
-
-/**
- * A struct to carry around Acts geometry on node tree, so as to not put 
- * all of the MakeActsGeometry tree
- */
-struct ActsTrackingGeometry{
-  ActsTrackingGeometry(){}
-  ActsTrackingGeometry(std::shared_ptr<const Acts::TrackingGeometry> tGeo,
-		       ActsExamples::Options::BFieldVariant mag,
-		       Acts::CalibrationContext calib,
-		       Acts::GeometryContext geoCtxt,
-		       Acts::MagneticFieldContext magFieldCtxt)
-  : tGeometry(tGeo)
-  , magField(mag)
-  , calibContext(calib)
-  , geoContext(geoCtxt)
-  , magFieldContext(magFieldCtxt)
-  {}
-  /// Tracking geometry and magnetic field, for fitter function
-  std::shared_ptr<const Acts::TrackingGeometry> tGeometry;
-  ActsExamples::Options::BFieldVariant magField;
-
-  /// Acts context, for Kalman options
-  Acts::CalibrationContext calibContext;
-  Acts::GeometryContext geoContext;
-  Acts::MagneticFieldContext magFieldContext;
-};
 
 /**
  * This class is responsible for creating Acts TrkrClusterSourceLinks from
